document and slightly restructure input
